fix(runtime): close runtime authority review gaps
Move peer response projection out of the runtime loop and into typed runtime input classification, remove the writable Agent LoopState shadow so public state is derived from turn authority snapshots, and harden wasm-pack installation in CI lanes that still used the raw rustwasm curl installer.
